requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh4f3eec9_6
docker-pycreds=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=h4a9f257_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.1.0=h9e868ea_3769
jpeg=9e=h5eee18b_1
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1u=h7f8727e_0
pathtools=0.1.2=pypi_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
protobuf=3.19.6=pypi_0
psutil=5.9.5=pypi_0
pyparsing=3.1.0=pypi_0
python=3.6.9=h265db76_0
pytorch=1.10.1=py3.6_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.27.1=pypi_0
scipy=1.5.4=pypi_0
sentry-sdk=1.25.1=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=58.0.4=py36h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.33.0=h62c20be_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.10.1=py36_cu102
torchvision=0.11.2=py36_cu102
tqdm=4.64.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.16=pypi_0
wandb=0.15.4=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.4.2=h5eee18b_0
zipp=3.6.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
